pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dawn-shuttle-novel"
version = "0.1.0"
description = "AI 小说生成器 - 完整的多 Agent 协作系统"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "dawn_shuttle_intelligence>=0.2.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
novel = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["RUF001", "RUF002", "RUF003"]
